Accessible iCarousel
Adding the Accessibility feature to iCarousel. This implementation is
based on the approached of making part of the UI not subject to
VoiceOver gestures. As a result VoiceOver doesn’t steal the iCarousel
single finger swipe. Delegate methods are used to process additional
Accessibility announcement upon swipe or single tap. Double tap is used
for item selection with VoiceOver on.